Fixed small item and scripting related bugs
Items now move from the inventory to their goal in the scene, not from
the starting point in the scene.
fadelayer now uses the correct opacity value instead of 0, though it
still does not actually fade the layer, but set's it to the given
opacity instantly.
There were also some minor adjustments to the scene (changed size of
hotspots; it still doesn't work how I want it!!!)